+++ hadoop-hdds/common/src/main/java/org/apache/hadoop/hdds/scm/ha/SCMHAUtils.java
@@ -22,7 +22,6 @@ package org.apache.hadoop.hdds.scm.ha;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
-import com.google.protobuf.ServiceException;
 import org.apache.hadoop.hdds.conf.ConfigurationException;
 import org.apache.hadoop.hdds.conf.ConfigurationSource;
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
@@ -34,8 +33,10 @@ import org.apache.hadoop.hdds.scm.pipeline.PipelineNotFoundException;
 import org.apache.hadoop.hdds.server.ServerUtils;
 import org.apache.hadoop.io.retry.RetryPolicy;
 import org.apache.hadoop.ipc.RemoteException;
+import org.apache.hadoop.ipc.RpcException;
 import org.apache.hadoop.ozone.ha.ConfUtils;
 import org.apache.hadoop.security.AccessControlException;
+import org.apache.hadoop.util.OzoneUtils;
 import org.apache.ratis.protocol.exceptions.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -311,8 +312,13 @@ public final class SCMHAUtils {
 
   public static RetryPolicy.RetryAction getRetryAction(int failovers, int retry,
       Exception e, int maxRetryCount, long retryInterval) {
-    // For AccessControl Exception where Client is not authenticated.
-    if (isAccessControlException(e)) {
+    Throwable unwrappedException = OzoneUtils.getUnwrappedException(e);
+    if (unwrappedException instanceof AccessControlException) {
+      // For AccessControl Exception where Client is not authenticated.
+      return RetryPolicy.RetryAction.FAIL;
+    } else if (OzoneUtils.shouldNotFailoverOnRpcException(unwrappedException)) {
+      // For some types of Rpc Exceptions, retrying on different server would
+      // not help.
       return RetryPolicy.RetryAction.FAIL;
     } else if (SCMHAUtils.checkRetriableWithNoFailoverException(e)) {
       if (retry < maxRetryCount) {
@@ -334,25 +340,5 @@ public final class SCMHAUtils {
         return RetryPolicy.RetryAction.FAIL;
       }
     }
-  }
-
-  /**
-   * Unwrap exception to check if it is some kind of access control problem.
-   * {@link AccessControlException}
-   */
-  public static boolean isAccessControlException(Exception ex) {
-    if (ex instanceof ServiceException) {
-      Throwable t = ex.getCause();
-      if (t instanceof RemoteException) {
-        t = ((RemoteException) t).unwrapRemoteException();
-      }
-      while (t != null) {
-        if (t instanceof AccessControlException) {
-          return true;
-        }
-        t = t.getCause();
-      }
-    }
-    return false;
   }
 }
